fix warning

+++ src/common/natto_map/src/map_converter.cpp
@@ -14,9 +14,6 @@
 
 #include "natto_map/map_converter.hpp"
 
-#include <algorithm>
-#include <cmath>
-
 namespace map_converter {
 
 map_converter::map_converter (const rclcpp::NodeOptions &node_options) : Node ("map_converter", node_options) {
@@ -62,7 +59,7 @@ void map_converter::map_callback (const natto_msgs::msg::Map::SharedPtr msg) {
     int width  = static_cast<int> ((max_x - min_x) / resolution_) + 1;
     int height = static_cast<int> ((max_y - min_y) / resolution_) + 1;
 
-    std::vector<int8_t> grid (width * height, 0);
+    std::vector<int8_t> grid (static_cast<size_t> (width * height), 0);
 
     auto world_to_grid = [&] (double x, double y, int &gx, int &gy) {
         gx = static_cast<int> ((x - min_x) / resolution_);
@@ -71,7 +68,7 @@ void map_converter::map_callback (const natto_msgs::msg::Map::SharedPtr msg) {
 
     auto set_cell = [&] (int gx, int gy) {
         if (gx >= 0 && gx < width && gy >= 0 && gy < height) {
-            grid[gy * width + gx] = 100;
+            grid[static_cast<size_t> (gy * width + gx)] = 100;
         }
     };
 
@@ -193,9 +190,9 @@ void map_converter::map_callback (const natto_msgs::msg::Map::SharedPtr msg) {
     nav_msgs::msg::OccupancyGrid occupancy_grid;
     occupancy_grid.header.stamp              = this->now ();
     occupancy_grid.header.frame_id           = "map";
-    occupancy_grid.info.resolution           = resolution_;
-    occupancy_grid.info.width                = width;
-    occupancy_grid.info.height               = height;
+    occupancy_grid.info.resolution           = static_cast<float> (resolution_);
+    occupancy_grid.info.width                = static_cast<uint32_t> (width);
+    occupancy_grid.info.height               = static_cast<uint32_t> (height);
     occupancy_grid.info.origin.position.x    = min_x;
     occupancy_grid.info.origin.position.y    = min_y;
     occupancy_grid.info.origin.position.z    = 0.0;
